archive/requirements_old.txt
# EDGE-QI Unified Requirements
# Generated: 2025-11-08 19:12:51
# Consolidated from multiple requirements.txt files

aiohttp==3.9.1
asyncpg==0.29.0
dash>=2.11.0
fastapi==0.109.0
numpy>=1.24.0
opencv-python>=4.8.0
paho-mqtt==1.6.1
pandas>=2.0.0
passlib[bcrypt]==1.7.4
pillow>=10.0.0
plotly>=5.15.0
psycopg2-binary==2.9.9
pydantic==2.5.3
pydantic-settings==2.1.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
python-socketio==5.11.0
redis==5.0.1
scikit-learn>=1.3.0
scipy>=1.11.0
sqlalchemy==2.0.25
streamlit>=1.25.0
uvicorn[standard]==0.27.0
